+++ boofuzz/sessions.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import
 
 import cPickle
 import datetime
-import itertools
 import logging
 import os
 import re
@@ -119,9 +118,9 @@ class Target(object):
         Returns:
             Received data.
         """
-        if max_bytes == None:
+        if max_bytes is None:
             max_bytes = self.max_recv_bytes
-        
+
         if self._fuzz_data_logger is not None:
             self._fuzz_data_logger.log_info("Receiving...")
 
@@ -279,6 +278,12 @@ class Session(pgraph.Graph):
         crash_threshold_request (int):  Maximum number of crashes allowed before a request is exhausted. Default 12.
         crash_threshold_element (int):  Maximum number of crashes allowed before an element is exhausted. Default 3.
         restart_sleep_time (int): Time in seconds to sleep when target can't be restarted. Default 5.
+        restart_callback (list of method): The registered method will be called after a failed post_test_case_callback
+                                           Default None.
+        pre_send_callback (list of method): The registered method will be called prior to each fuzz request.
+                                            Default None.
+        post_test_case_callback (list of method): The registered method will be called after each fuzz test case.
+                                                  Default None.
         web_port (int):         Port for monitoring fuzzing campaign via a web browser. Default 26000.
         fuzz_data_logger (fuzz_logger.FuzzLogger): DEPRECATED. Use fuzz_loggers instead.
         fuzz_loggers (list of ifuzz_logger.IFuzzLogger): For saving test data and results.. Default Log to STDOUT.
@@ -316,6 +321,9 @@ class Session(pgraph.Graph):
                  crash_threshold_request=12,
                  crash_threshold_element=3,
                  restart_sleep_time=5,
+                 restart_callback=None,
+                 pre_send_callback=None,
+                 post_test_case_callback=None,
                  fuzz_data_logger=None,
                  fuzz_loggers=None,
                  fuzz_db_keep_only_n_pass_cases=0,
@@ -363,7 +371,9 @@ class Session(pgraph.Graph):
         self._receive_data_after_fuzz = receive_data_after_fuzz
         self._skip_current_node_after_current_test_case = False
         self._skip_current_element_after_current_test_case = False
-        self._post_test_case_methods = []
+        self._pre_send_methods = pre_send_callback
+        self._post_test_case_methods = post_test_case_callback
+        self._restart_methods = restart_callback
 
         self.web_interface_thread = self.build_webapp_thread(port=self.web_port)
 
@@ -660,6 +670,9 @@ class Session(pgraph.Graph):
             self._fuzz_data_logger.log_error("Restarting the target failed, exiting.")
             self.export_file()
             raise
+        except exception.BoofuzzCallbackError:
+            self._fuzz_data_logger.log_error("Custom callback function raised uncaught exception, exiting.")
+            self.export_file()
         except exception.BoofuzzTargetConnectionFailedError:
             # exception should have already been handled but rethrown in order to escape test run
             pass
@@ -854,6 +867,9 @@ class Session(pgraph.Graph):
         Args:
             method (function): A method with the same parameters as :func:`~Session.post_send`
             """
+        if self._post_test_case_methods is None:
+            self._post_test_case_methods = []
+
         self._post_test_case_methods.append(method)
 
     # noinspection PyUnusedLocal
@@ -869,7 +885,6 @@ class Session(pgraph.Graph):
             session (Session): Session object calling post_send.
                 Useful properties include last_send and last_recv.
 
-            sock: DEPRECATED Included for backward-compatibility. Same as target.
             args: Implementations should include \*args and \**kwargs for forward-compatibility.
             kwargs: Implementations should include \*args and \**kwargs for forward-compatibility.
         """
@@ -877,10 +892,9 @@ class Session(pgraph.Graph):
         self._fuzz_data_logger.log_info("No post_send callback registered.")
 
     # noinspection PyMethodMayBeStatic
-    def pre_send(self, sock):
+    def pre_send(self, target):
         """
-        Overload or replace this routine to specify actions to run prior to each fuzz request. The order of events is
-        as follows::
+        Execute custom methods to run prior to each fuzz request. The order of events is as follows::
 
             pre_send() - req - callback ... req - callback - post_send()
 
@@ -889,47 +903,75 @@ class Session(pgraph.Graph):
         @see: pre_send()
 
         Args:
-            sock (Socket): Connected socket to target
+            target (session.target): Target we are sending data to
         """
 
-        # default to doing nothing.
-        pass
+        if self._pre_send_methods is not None:
+            try:
+                for f in self._pre_send_methods:
+                    self._fuzz_data_logger.open_test_step('Pre_Send callback: "{0}"'.format(f.__name__))
+                    f(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self, sock=target)
+            except Exception:
+                self._fuzz_data_logger.log_error(constants.ERR_CALLBACK_FUNC.format(func_name="pre_send")
+                                                 + traceback.format_exc())
+                raise exception.BoofuzzCallbackError()
+            finally:
+                self._fuzz_data_logger.open_test_step('Cleaning up connections from callbacks')
+                target.close()
+        else:
+            pass
 
     def restart_target(self, target):
         """
         Restart the fuzz target. If a VMControl is available revert the snapshot, if a process monitor is available
-        restart the target process. Otherwise, do nothing.
+        restart the target process. If custom restart methods are registered, execute them. Otherwise, do nothing.
 
         Args:
             target (session.target): Target we are restarting
 
-        @raise sex.BoofuzzRestartFailedError if restart fails.
+        @raise exception.BoofuzzRestartFailedError if restart fails.
         """
 
-        self._fuzz_data_logger.open_test_step("restarting target")
+        self._fuzz_data_logger.open_test_step("Restarting target")
         if len(self.on_failure) > 0:
             for f in self.on_failure:
-                self._fuzz_data_logger.open_test_step("calling registered on_failure method")
+                self._fuzz_data_logger.open_test_step("Calling registered on_failure method")
                 f(logger=self._fuzz_data_logger)
         # vm restarting is the preferred method so try that before procmon.
         elif target.vmcontrol:
-            self._fuzz_data_logger.log_info("restarting target virtual machine")
+            self._fuzz_data_logger.log_info("Restarting target virtual machine")
             target.vmcontrol.restart_target()
 
         # if we have a connected process monitor, restart the target process.
         elif target.procmon:
-            self._fuzz_data_logger.log_info("restarting target process")
+            self._fuzz_data_logger.log_info("Restarting target process")
 
             if not target.procmon.restart_target():
                 raise exception.BoofuzzRestartFailedError()
 
-            self._fuzz_data_logger.log_info("giving the process 3 seconds to settle in ")
+            self._fuzz_data_logger.log_info("Giving the process 3 seconds to settle in")
             time.sleep(3)
+
+        # if we have custom restart methods, execute them
+        elif self._restart_methods is not None:
+            try:
+                for f in self._restart_methods:
+                    self._fuzz_data_logger.open_test_step('Target restart callback: "{0}"'.format(f.__name__))
+                    f(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self, sock=target)
+            except exception.BoofuzzRestartFailedError:
+                raise
+            except Exception:
+                self._fuzz_data_logger.log_error(constants.ERR_CALLBACK_FUNC.format(func_name="restart_target")
+                                                 + traceback.format_exc())
+                raise exception.BoofuzzCallbackError()
+            finally:
+                self._fuzz_data_logger.open_test_step('Cleaning up connections from callbacks')
+                target.close()
 
         # otherwise all we can do is wait a while for the target to recover on its own.
         else:
             self._fuzz_data_logger.log_info(
-                "no reset handler available... sleeping for %d seconds" % self.restart_sleep_time
+                "No reset handler available... sleeping for {} seconds".format(self.restart_sleep_time)
             )
             time.sleep(self.restart_sleep_time)
 
@@ -1363,13 +1405,9 @@ class Session(pgraph.Graph):
         return "{0}.{1}.{2}".format(message_path, primitive_under_test, self.fuzz_node.mutant_index)
 
     def _post_send(self, target):
-        try:
-            deprecated_callbacks = [self.post_send]
-        except AttributeError:
-            deprecated_callbacks = []
-        if len(self._post_test_case_methods) + len(deprecated_callbacks) > 0:
+        if self._post_test_case_methods is not None:
             try:
-                for f in itertools.chain(self._post_test_case_methods, deprecated_callbacks):
+                for f in self._post_test_case_methods:
                     self._fuzz_data_logger.open_test_step('Post- test case callback: "{0}"'.format(f.__name__))
                     f(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self, sock=target)
             except exception.BoofuzzTargetConnectionReset:
@@ -1381,8 +1419,9 @@ class Session(pgraph.Graph):
             except exception.BoofuzzTargetConnectionFailedError:
                 self._fuzz_data_logger.log_fail(constants.ERR_CONN_FAILED)
             except Exception:
-                self._fuzz_data_logger.log_fail(
-                    "Custom post_send method raised uncaught Exception." + traceback.format_exc())
+                self._fuzz_data_logger.log_error(constants.ERR_CALLBACK_FUNC.format(func_name="post_send")
+                                                 + traceback.format_exc())
+                raise exception.BoofuzzCallbackError()
             finally:
                 self._fuzz_data_logger.open_test_step('Cleaning up connections from callbacks')
                 target.close()